requirements.txt
matplotlib==3.7.1
numpy==1.22.4
Pillow==9.5.0
torch==2.0.1
tqdm==4.61.1
